+++ core/CHANGELOG.md
@@ -8,7 +8,7 @@ and this project adheres to
 
 ## Unreleased
 
-### Added
+### 0.14.0
 
 - `addToLabelIndex`: a function to include a thing in a label index, so that it can be found via search
 - `LabelIndex.contains`: a method to check if a label index contains a specific URI